Some work for fMRI analysis.
