+++ tetrad-lib/src/main/java/edu/cmu/tetrad/graph/GraphUtils.java
@@ -2842,117 +2842,7 @@ public final class GraphUtils {
         return b;
     }
 
-    /**
-     * Builds a data set with columns being the set of all edges in the supplied graphs
-     * plus a class label column for which group they belong to.
-     */
-    public static void printEdgeDataSet(List<List<Graph>> graphs, String path, String prefix) {
-        File file1 = new File(path, prefix + ".data.txt");
-        File file2 = new File(path, prefix + ".edges.txt");
-        File dir = new File(path);
 
-        PrintStream out1 = null;
-        PrintStream out2 = null;
-        try {
-            out1 = new PrintStream(new FileOutputStream(file1));
-            out2 = new PrintStream(new FileOutputStream(file2));
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-            throw new RuntimeException();
-        }
-
-        Set<Node> _nodes = new HashSet<>();
-
-        for (List<Graph> _graphs : graphs) {
-            for (Graph graph : _graphs) {
-                for (Node node : graph.getNodes()) {
-                    boolean found = false;
-
-                    for (Node _node : _nodes) {
-                        if (_node.getName().equals(node.getName())) {
-                            found = true;
-                        }
-                    }
-
-                    if (!found) {
-                        _nodes.add(node);
-                    }
-                }
-            }
-        }
-
-        List<List<Graph>> graphs2 = new ArrayList<>();
-
-        for (List<Graph> _graphs : graphs) {
-            List<Graph> _graphs2 = new ArrayList<>();
-
-            for (Graph graph : _graphs) {
-                Graph graph2 = GraphUtils.replaceNodes(graph, new ArrayList<Node>(_nodes));
-                _graphs2.add(graph2);
-            }
-
-            graphs2.add(_graphs2);
-        }
-
-        Set<Edge> _edges = new HashSet<>();
-
-        for (List<Graph> _graphs : graphs) {
-            for (Graph graph : _graphs) {
-                _edges.addAll(graph.getEdges());
-            }
-        }
-
-        List<Edge> __edges = new ArrayList<>();
-        System.out.println("# edges = " + _edges.size());
-
-        for (Edge edge : _edges) {
-            int count = 0;
-            int total = 0;
-
-            for (List<Graph> _graphs : graphs) {
-                for (Graph graph : _graphs) {
-                    if (graph.containsEdge(edge)) {
-                        count++;
-                    }
-
-                    total++;
-                }
-            }
-
-            if (count >= 2 && count <= 10) {
-                __edges.add(edge);
-            }
-        }
-
-        System.out.println("# __edges = " + __edges.size());
-
-        for (int i = 0; i < __edges.size(); i++) {
-//            out1.print(edges.get(i) + "\t");
-            out1.print("X" + (i + 1) + "\t");
-        }
-
-        out1.println("Group");
-
-        for (int i = 0; i < graphs2.size(); i++) {
-            for (Graph graph : graphs2.get(i)) {
-                for (Edge edge : __edges) {
-                    out1.print(graph.containsEdge(edge) ? "1\t" : "0\t");
-                }
-
-                out1.println(i);
-            }
-        }
-
-        out1.close();
-
-        for (int i = 0; i < __edges.size(); i++) {
-            out2.println("X" + (i + 1) + "\t" + __edges.get(i));
-        }
-
-        out2.println("Group");
-
-        out2.close();
-    }
 
     private static StringBuilder directedEdges(List<Graph> directedGraphs) {
         Set<Edge> directedEdgesSet = new HashSet<>();
